+++ R/cnd-abort.R
@@ -62,9 +62,10 @@
 #' @inheritParams cnd
 #' @param message The message to display.
 #'
-#'   Experimental: Can also be a named character vector, in which case
-#'   the message is assembled as a list of bullets. See
-#'   [cnd_message()] to learn how names control the bulleted output.
+#'   If not supplied, it is expected that the message is generated
+#'   lazily through [conditionMessage()][cnd_message]. In that case,
+#'   `class` must be supplied. Only `inform()` allows empty messages
+#'   as it is occasionally useful to build user output incrementally.
 #' @param class Subclass of the condition. This allows your users
 #'   to selectively handle the conditions signalled by your functions.
 #' @param ... Additional data to be stored in the condition object.
@@ -132,7 +133,7 @@
 #'
 #' }
 #' @export
-abort <- function(message = "",
+abort <- function(message = NULL,
                   class = NULL,
                   ...,
                   trace = NULL,
@@ -155,6 +156,7 @@ abort <- function(message = "",
     trace <- trace_trim_context(trace, context)
   }
 
+  message <- validate_signal_message(message, class)
   message <- collapse_cnd_message(message)
 
   cnd <- error_cnd(class,